Set Git tracking tags outside of an MLFlow Project

This is a utility to make it easier to set tracking tags. Basically, it tricks MLFlow into thinking the code is running from an MLFlow Project when it's not, so you get nice Git tracking in the UI without needing an MLProject file or to clone a repo to get the same benefits.

set_git_tracking_tags(
  run_id = get_active_run_id(),
  client = mlflow_client(),
  git_repo_url = Sys.getenv("GIT_REPO_URL"),
  git_repo_subdir,
  source_git_commit = system("git rev-parse HEAD", intern = TRUE),
  source_git_branch = system("git rev-parse --abbrev-ref HEAD", intern = TRUE),
  project_entrypoint = "main",
  project_backend = "local"
)

Arguments

run_id

An MLFlow run ID. Auto-generated if missing

client

An MLFlow client. Auto-generated if missing

git_repo_url

A URL for the Git repo where the code is being run from This is expected to be of the form https://<<GITHUB_TOKEN>>@github.com/<<org>>/<<repo>>

git_repo_subdir

The subdirectory in the repo where your code lives

source_git_commit

A git commit hash to tag a run with. Defaults to git2r::revparse_single(".", revision = "HEAD")$sha

source_git_branch

The git branch the code is running on. Defaults to system("git rev-parse --abbrev-ref HEAD", intern = TRUE) (i.e. the current branch)

project_entrypoint

An entrypoint for the project. Defaults to "main"

project_backend

A backend for the project. Defaults to "local"

Value

Returns a list of tags for the run_id

Details

This also makes it easier to have MLFlow play nicely with renv.